enh(broker): big metrics do not fail anymore when inserted in database
* enh(tests): big metrics do not fail anymore when inserted in database
* fix(broker): new script to compute database columns max sizes
* fix(broker/bam): the good size of float is inserted into the database for bam
* fix(broker): perfdata class contains float numbers now.
* enh(tests): big metrics do not fail anymore when inserted in database

REFS: MON-116619

+++ broker/unified_sql/test/perfdata.cc
@@ -240,9 +240,9 @@ TEST_F(UnifiedSqlParserParsePerfdata, Simple2) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(18.0);
   expected.unit("%");
-  expected.warning(std::numeric_limits<double>::infinity());
+  expected.warning(std::numeric_limits<float>::infinity());
   expected.warning_low(15.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(10.0);
   expected.min(0.0);
   expected.max(100.0);
@@ -266,7 +266,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(2.45698);
   expected.unit("s");
-  expected.max(std::numeric_limits<double>::infinity());
+  expected.max(std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ++it;
 
@@ -278,7 +278,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.unit("B/s");
   expected.warning(5.0);
   expected.warning_low(0.0);
-  expected.min(-std::numeric_limits<double>::infinity());
+  expected.min(-std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ++it;
 
@@ -298,7 +298,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.value(1234.0);
   expected.warning(10.0);
   expected.warning_low(0.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(11.0);
   ASSERT_TRUE(expected == *it);
   ++it;
@@ -309,7 +309,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::counter);
   expected.value(1234.0);
   expected.warning(10.0);
-  expected.warning_low(-std::numeric_limits<double>::infinity());
+  expected.warning_low(-std::numeric_limits<float>::infinity());
   expected.critical(30.0);
   expected.critical_low(20.0);
   ASSERT_TRUE(expected == *it);
@@ -332,10 +332,10 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex1) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(9.0);
   expected.unit("queries_per_second");
-  expected.warning(std::numeric_limits<double>::infinity());
+  expected.warning(std::numeric_limits<float>::infinity());
   expected.warning_low(10.0);
   expected.warning_mode(true);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(5.0);
   expected.critical_mode(true);
   expected.min(0.0);
@@ -447,7 +447,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex2) {
   expected.value_type(misc::perfdata::gauge);
   expected.value(2.45698);
   expected.unit("s");
-  expected.max(std::numeric_limits<double>::infinity());
+  expected.max(std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
   ++it;
@@ -460,7 +460,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex2) {
   expected.unit("B/s");
   expected.warning(5.0);
   expected.warning_low(0.0);
-  expected.min(-std::numeric_limits<double>::infinity());
+  expected.min(-std::numeric_limits<float>::infinity());
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
   ++it;
@@ -492,7 +492,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex2) {
   expected.value(1234.17);
   expected.warning(10.0);
   expected.warning_low(0.0);
-  expected.critical(std::numeric_limits<double>::infinity());
+  expected.critical(std::numeric_limits<float>::infinity());
   expected.critical_low(11.0);
   ASSERT_TRUE(expected == *it);
   ASSERT_FALSE(expected != *it);
@@ -504,7 +504,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, Complex2) {
   expected.value_type(misc::perfdata::counter);
   expected.value(1234.147);
   expected.warning(10.0);
-  expected.warning_low(-std::numeric_limits<double>::infinity());
+  expected.warning_low(-std::numeric_limits<float>::infinity());
   expected.critical(30.0);
   expected.critical_low(20.0);
   ASSERT_TRUE(expected == *it);
@@ -546,7 +546,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, BadMetric) {
   int i = 1;
   for (auto& p : lst) {
     ASSERT_EQ(p.name(), fmt::format("user{}", i));
-    ASSERT_EQ(p.value(), static_cast<double>(i));
+    ASSERT_EQ(p.value(), static_cast<float>(i));
     ++i;
   }
 }
@@ -559,7 +559,7 @@ TEST_F(UnifiedSqlParserParsePerfdata, BadMetric1) {
   int i = 1;
   for (auto& p : lst) {
     ASSERT_EQ(p.name(), fmt::format("user{}", i));
-    ASSERT_EQ(p.value(), static_cast<double>(i));
+    ASSERT_EQ(p.value(), static_cast<float>(i));
     ++i;
   }
 }
